+++ changelog.md
@@ -2,6 +2,7 @@
 
 * [Bug] - Fix commit `3562a62` that was wiped out by release 2.1.7
 * [Chore] - Update dependencies
+* [Bug] - Fix NightBloom File ID extraction regex following API change
 
 ### Version 2.1.7